[EI-153] [web] refactor: Refactor Icon Button
