+++ web/app/ui/components/view/atom/editable-text.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { useEffect, useRef, useState } from 'react';
-import IconButton from './icon-button/icon-button';
+import IconButton from './icons/icon-button';
 import { XCircleIcon } from '@heroicons/react/solid';
 import { useDebouncedCallback } from 'use-debounce';
 import { twMerge } from 'tailwind-merge';
